ci: optimize and condense E2E tests (#1263)
* First pass; combine non-output receipt tests into one

* Combine explorer-related tests

* Combine tests that hit /ping

* Combine receipt tests again

* Finish optimization of e2e::indexing_postgres

* Remove unused replace-indexer tests from indexing_postgres

* Remove outdated, unused test from integration::service

* Condense GraphQL API E2E tests

* Use get_or_create() in test indexer

* Remove IDs checks in E2E tests

* Cleanup formatting and folder structure

* Move http_client into WebTestComponents

* Upgrade Sway version

* cargo fmt

* Adjust EXPECTED_CONTRACT_ID

* Upgrade fuels-rs to v0.45.1
