+++ .github/workflows/ci.yml
@@ -264,7 +264,7 @@ jobs:
           bash scripts/stripper.bash simple_wasm.wasm
           cp simple_wasm.wasm target/wasm32-unknown-unknown/release/
       - name: Run E2E and integration tests
-        run: cargo test --locked --all-targets --features e2e,postgres -p
+        run: cargo test --locked --all-targets --features postgres -p
           fuel-indexer-tests -- --test-threads=1
   build-default-indexer:
     if: needs.set-env-vars.outputs.IS_RELEASE_OR_MASTER_OR_DEVELOP_OR_SEMVER != 'true'
