+++ packages/fuel-indexer-tests/components/indices/fuel-indexer-test/src/lib.rs
@@ -8,48 +8,30 @@ use fuel_indexer_utils::prelude::*;
 mod fuel_indexer_test {
 
     fn fuel_indexer_test_blocks(block_data: BlockData) {
-        let block = BlockEntity {
-            id: id8(block_data.id),
-            height: block_data.height,
-            timestamp: block_data.time,
-        };
-
-        block.save();
+        let block = BlockEntity::new(block_data.height, block_data.time).get_or_create();
 
         let input_data = r#"{"foo":"bar"}"#.to_string();
 
-        for tx in block_data.transactions.iter() {
-            let tx = TxEntity {
-                id: id8(tx.id),
-                block: block.id,
-                timestamp: block_data.time,
-                input_data: Json(input_data.clone()),
-            };
-            tx.save();
+        for _tx in block_data.transactions.iter() {
+            TxEntity::new(block.id, Json(input_data.clone()), block_data.time)
+                .get_or_create();
         }
     }
 
     fn fuel_indexer_test_ping(ping: Ping) {
         info!("fuel_indexer_test_ping handling a Ping event: {:?}.", ping);
 
-        let entity = PingEntity {
-            id: ping.id,
-            value: ping.value,
-            message: ping.message.to_string(),
-        };
-
-        entity.save();
+        PingEntity::new(ping.value, ping.message.to_string()).get_or_create();
     }
 
     fn fuel_indexer_test_u16(_ping: Ping) {
         info!("fuel_indexer_test_ping handling a U16 event.");
-        let u16entity = U16Entity {
-            id: 9999,
-            value1: 340282366920938463463374607431768211454, // 2**128-2
-            value2: 170141183460469231731687303715884105727, // 2**127-1
-        };
 
-        u16entity.save();
+        U16Entity::new(
+            340282366920938463463374607431768211454,
+            170141183460469231731687303715884105727,
+        )
+        .get_or_create();
     }
 
     fn fuel_indexer_test_transfer(transfer: Transfer) {
@@ -66,15 +48,7 @@ mod fuel_indexer_test {
             ..
         } = transfer;
 
-        let entity = TransferEntity {
-            id: id8(contract_id),
-            contract_id,
-            recipient: to,
-            amount,
-            asset_id,
-        };
-
-        entity.save();
+        TransferEntity::new(contract_id, to, amount, asset_id).get_or_create();
     }
 
     fn fuel_indexer_test_transferout(transferout: TransferOut) {
@@ -91,15 +65,7 @@ mod fuel_indexer_test {
             ..
         } = transferout;
 
-        let entity = TransferOutEntity {
-            id: id8(contract_id),
-            contract_id,
-            recipient: to,
-            amount,
-            asset_id,
-        };
-
-        entity.save();
+        TransferOutEntity::new(contract_id, to, amount, asset_id).get_or_create();
     }
 
     fn fuel_indexer_test_log(log: Log) {
@@ -112,27 +78,18 @@ mod fuel_indexer_test {
             ..
         } = log;
 
-        let entity = LogEntity {
-            id: id8(contract_id),
-            contract_id: log.contract_id,
-            ra,
-            rb,
-        };
-
-        entity.save();
+        LogEntity::new(contract_id, ra, rb).get_or_create();
     }
 
     fn fuel_indexer_test_logdata(logdata_entity: Pung) {
         info!("fuel_indexer_test_logdata handling LogData event.");
 
-        let entity = PungEntity {
-            id: logdata_entity.id,
-            value: logdata_entity.value,
-            is_pung: logdata_entity.is_pung,
-            pung_from: logdata_entity.pung_from,
-        };
-
-        entity.save();
+        PungEntity::new(
+            logdata_entity.value,
+            logdata_entity.is_pung,
+            logdata_entity.pung_from,
+        )
+        .get_or_create();
     }
 
     fn fuel_indexer_test_scriptresult(scriptresult: ScriptResult) {
@@ -140,25 +97,13 @@ mod fuel_indexer_test {
 
         let ScriptResult { result, gas_used } = scriptresult;
 
-        let entity = ScriptResultEntity {
-            id: result,
-            result,
-            gas_used,
-            blob: vec![1u8, 1, 1, 1, 1].into(),
-        };
-
-        entity.save();
+        ScriptResultEntity::new(result, gas_used, vec![1u8; 5].into()).get_or_create();
     }
 
     fn fuel_indexer_test_messageout_data(example_message: ExampleMessageStruct) {
         info!("fuel_indexer_test_messageout_data handling MessageOut data event");
 
-        let entity = MessageEntity {
-            id: example_message.id,
-            message: example_message.message.to_string(),
-        };
-
-        entity.save();
+        MessageEntity::new(example_message.message.to_string()).get_or_create();
     }
 
     fn fuel_indexer_test_messageout(messageout: MessageOut) {
@@ -175,31 +120,15 @@ mod fuel_indexer_test {
             ..
         } = messageout;
 
-        let entity = MessageOutEntity {
-            id: id8(message_id),
-            message_id,
-            sender,
-            recipient,
-            amount,
-            nonce,
-            len,
-            digest,
-        };
-
-        entity.save();
+        MessageOutEntity::new(message_id, sender, recipient, amount, nonce, len, digest)
+            .get_or_create();
     }
 
     fn fuel_indexer_test_callreturn(pungentity: Pung) {
         info!("fuel_indexer_test_callreturn handling Pung event.");
 
-        let entity = PungEntity {
-            id: pungentity.id,
-            value: pungentity.value,
-            is_pung: pungentity.is_pung,
-            pung_from: pungentity.pung_from,
-        };
-
-        entity.save();
+        PungEntity::new(pungentity.value, pungentity.is_pung, pungentity.pung_from)
+            .get_or_create();
     }
 
     fn fuel_indexer_test_multiargs(
@@ -212,44 +141,20 @@ mod fuel_indexer_test {
             "fuel_indexer_test_multiargs handling Pung, Pong, Ping, and BlockData events."
         );
 
-        let block = BlockEntity {
-            id: id8(block_data.id),
-            height: block_data.height,
-            timestamp: block_data.time,
-        };
+        BlockEntity::new(block_data.height, block_data.time).get_or_create();
 
-        block.save();
+        PungEntity::new(pung.value, pung.is_pung, pung.pung_from).get_or_create();
 
-        let pu = PungEntity {
-            id: pung.id,
-            value: pung.value,
-            is_pung: pung.is_pung,
-            pung_from: pung.pung_from,
-        };
+        PongEntity::new(pong.value).get_or_create();
 
-        pu.save();
-
-        let po = PongEntity {
-            id: pong.id,
-            value: pong.value,
-        };
-
-        po.save();
-
-        let pi = PingEntity {
-            id: ping.id,
-            value: ping.value,
-            message: ping.message.to_string(),
-        };
-
-        pi.save();
+        PingEntity::new(ping.value, ping.message.to_string()).get_or_create();
     }
 
-    fn fuel_indexer_test_optional_schema_fields(optional: Ping) {
+    fn fuel_indexer_test_optional_schema_fields(_block: BlockData) {
         info!("fuel_indexer_test_optional_schema_fields handling Ping event and setting optional fields.");
 
         let entity = OptionEntity {
-            id: optional.id,
+            id: 8675309,
             int_required: 100,
             int_optional_some: Some(999),
             addr_optional_none: None,
@@ -264,13 +169,13 @@ mod fuel_indexer_test {
     ) {
         info!("fuel_indexer_test_tuple handling ComplexTupleStruct and SimpleTupleStruct events.");
         let data: (u32, (u64, bool, (SizedAsciiString<5>, TupleStructItem))) = event.data;
-        let entity = TupleEntity {
-            id: data.1 .0,
-            complex_a: data.1 .2 .0.to_string(),
-            complex_b: data.1 .2 .1.id,
-            simple_a: logdata_entity.data.2.to_string(),
-        };
-        entity.save();
+
+        TupleEntity::new(
+            data.1 .2 .0.to_string(),
+            data.1 .2 .1.id,
+            logdata_entity.data.2.to_string(),
+        )
+        .get_or_create();
     }
 
     fn fuel_indexer_test_pure_function(call: Call) {
@@ -285,232 +190,78 @@ mod fuel_indexer_test {
             amount,
         } = call;
 
-        let entity = CallEntity {
-            id: 123,
-            contract_id,
-            callee: to,
-            asset_id,
-            gas,
-            fn_name,
-            amount,
-        };
-
-        entity.save();
+        CallEntity::new(contract_id, to, asset_id, gas, fn_name, amount).get_or_create();
     }
 
     fn fuel_indexer_test_deeply_nested_schema_fields(_deeply_nested: SimpleQueryStruct) {
         info!("fuel_indexer_test_deeply_nested_schema_fields handling DeeplyNestedQueryTestStruct event.");
 
-        let genre1 = Genre {
-            id: 1,
-            name: "horror".to_string(),
-        };
-        let genre2 = Genre {
-            id: 2,
-            name: "mystery".to_string(),
-        };
-        let genre3 = Genre {
-            id: 3,
-            name: "business".to_string(),
-        };
+        let genre1 = Genre::new("horror".to_string()).get_or_create();
+        let genre2 = Genre::new("mystery".to_string()).get_or_create();
+        let genre3 = Genre::new("business".to_string()).get_or_create();
 
-        genre1.save();
-        genre2.save();
-        genre3.save();
+        let author1 = Author::new("Brian".to_string(), genre1.id).get_or_create();
+        let author2 = Author::new("James".to_string(), genre2.id).get_or_create();
+        let author3 = Author::new("Susan".to_string(), genre3.id).get_or_create();
 
-        let person1 = Person {
-            id: 1,
-            name: "Ava".to_string(),
-        };
-        let person2 = Person {
-            id: 2,
-            name: "Noel".to_string(),
-        };
-        let person3 = Person {
-            id: 3,
-            name: "Julio".to_string(),
-        };
+        let person1 = Person::new("Ava".to_string()).get_or_create();
+        let person2 = Person::new("Noel".to_string()).get_or_create();
+        let person3 = Person::new("Julio".to_string()).get_or_create();
 
-        person1.save();
-        person2.save();
-        person3.save();
+        let planet = Planet::new("Earth".to_string()).get_or_create();
 
-        let planet = Planet {
-            id: 1,
-            name: "Earth".to_string(),
-        };
+        let continent1 =
+            Continent::new("North America".to_string(), planet.id).get_or_create();
+        let continent2 = Continent::new("Europe".to_string(), planet.id).get_or_create();
 
-        planet.save();
+        let country1 = Country::new("Jamaica".to_string(), continent1.id).get_or_create();
+        let country2 = Country::new("France".to_string(), continent2.id).get_or_create();
+        let country3 = Country::new("Ireland".to_string(), continent2.id).get_or_create();
 
-        let continent1 = Continent {
-            id: 1,
-            name: "North America".to_string(),
-            planet: planet.id,
-        };
+        let region1 =
+            Region::new("Westmoreland".to_string(), country1.id).get_or_create();
+        let region2 = Region::new("Corsica".to_string(), country2.id).get_or_create();
+        let region3 = Region::new("Donegal".to_string(), country3.id).get_or_create();
 
-        let continent2 = Continent {
-            id: 2,
-            name: "Europe".to_string(),
-            planet: planet.id,
-        };
+        let city1 = City::new("Savanna-la-Mar".to_string(), region1.id).get_or_create();
+        let city2 = City::new("Ajaccio".to_string(), region2.id).get_or_create();
+        let city3 = City::new("Letterkenny".to_string(), region3.id).get_or_create();
 
-        continent1.save();
-        continent2.save();
+        let library1 =
+            Library::new("Scholar Library".to_string(), city1.id).get_or_create();
+        let library2 =
+            Library::new("Ronoke Library".to_string(), city2.id).get_or_create();
+        let library3 =
+            Library::new("Scholastic Library".to_string(), city3.id).get_or_create();
 
-        let country1 = Country {
-            id: 1,
-            name: "Jamaica".to_string(),
-            continent: continent1.id,
-        };
+        let book1 = Book::new(
+            "Gone with the Wind".to_string(),
+            author1.id,
+            library1.id,
+            genre1.id,
+        )
+        .get_or_create();
 
-        let country2 = Country {
-            id: 2,
-            name: "France".to_string(),
-            continent: continent2.id,
-        };
+        let book2 = Book::new("Othello".to_string(), author2.id, library2.id, genre2.id)
+            .get_or_create();
 
-        let country3 = Country {
-            id: 3,
-            name: "Ireland".to_string(),
-            continent: continent2.id,
-        };
+        let book3 = Book::new(
+            "Cyberpunk 2021".to_string(),
+            author3.id,
+            library3.id,
+            genre3.id,
+        )
+        .get_or_create();
 
-        country1.save();
-        country2.save();
-        country3.save();
+        let sponsor1 =
+            Sponsor::new("Fuel Labs".to_string(), 100, person1.id).get_or_create();
 
-        let region1 = Region {
-            id: 1,
-            name: "Westmoreland".to_string(),
-            country: country1.id,
-        };
+        let sponsor2 = Sponsor::new("Fuel Labs, Part Two".to_string(), 200, person2.id)
+            .get_or_create();
 
-        let region2 = Region {
-            id: 2,
-            name: "Corsica".to_string(),
-            country: country2.id,
-        };
-
-        let region3 = Region {
-            id: 3,
-            name: "Donegal".to_string(),
-            country: country3.id,
-        };
-
-        region1.save();
-        region2.save();
-        region3.save();
-
-        let city1 = City {
-            id: 1,
-            name: "Savanna-la-Mar".to_string(),
-            region: region1.id,
-        };
-        let city2 = City {
-            id: 2,
-            name: "Ajaccio".to_string(),
-            region: region2.id,
-        };
-        let city3 = City {
-            id: 3,
-            name: "Letterkenny".to_string(),
-            region: region3.id,
-        };
-
-        city1.save();
-        city2.save();
-        city3.save();
-
-        let author1 = Author {
-            id: 1,
-            name: "Brian".to_string(),
-            genre: genre1.id,
-        };
-        let author2 = Author {
-            id: 2,
-            name: "James".to_string(),
-            genre: genre2.id,
-        };
-        let author3 = Author {
-            id: 3,
-            name: "Susan".to_string(),
-            genre: genre3.id,
-        };
-
-        author1.save();
-        author2.save();
-        author3.save();
-
-        let library1 = Library {
-            id: 1,
-            name: "Scholar Library".to_string(),
-            city: city1.id,
-        };
-        let library2 = Library {
-            id: 2,
-            name: "Ronoke Library".to_string(),
-            city: city2.id,
-        };
-        let library3 = Library {
-            id: 3,
-            name: "Scholastic Library".to_string(),
-            city: city3.id,
-        };
-
-        library1.save();
-        library2.save();
-        library3.save();
-
-        let book1 = Book {
-            id: 1,
-            name: "Gone with the Wind".to_string(),
-            author: author1.id,
-            library: library1.id,
-            genre: genre1.id,
-        };
-        let book2 = Book {
-            id: 2,
-            name: "Othello".to_string(),
-            author: author2.id,
-            library: library2.id,
-            genre: genre2.id,
-        };
-        let book3 = Book {
-            id: 3,
-            name: "Cyberpunk 2021".to_string(),
-            author: author3.id,
-            library: library3.id,
-            genre: genre3.id,
-        };
-
-        book1.save();
-        book2.save();
-        book3.save();
-
-        let sponsor1 = Sponsor {
-            id: 1,
-            name: "Fuel Labs".to_string(),
-            amount: 100,
-            representative: person1.id,
-        };
-
-        let sponsor2 = Sponsor {
-            id: 2,
-            name: "Fuel Labs, Part Two".to_string(),
-            amount: 200,
-            representative: person2.id,
-        };
-
-        let sponsor3 = Sponsor {
-            id: 3,
-            name: "Fuel Labs, Yet A Third".to_string(),
-            amount: 300,
-            representative: person3.id,
-        };
-
-        sponsor1.save();
-        sponsor2.save();
-        sponsor3.save();
+        let sponsor3 =
+            Sponsor::new("Fuel Labs, Yet A Third".to_string(), 300, person3.id)
+                .get_or_create();
 
         let bookclub1 = BookClub {
             id: 1,
@@ -539,24 +290,14 @@ mod fuel_indexer_test {
     }
 
     fn fuel_indexer_test_nested_query_explicit_foreign_keys_schema_fields(
-        explicit: ExplicitQueryStruct,
+        _block: BlockData,
     ) {
         info!("fuel_indexer_test_nested_query_explicit_foreign_keys_schema_fields handling ExplicitQueryTestStruct event.");
 
-        let municipality = Municipality {
-            id: explicit.id,
-            name: "Republic of Indexia".to_string(),
-        };
+        let municipality =
+            Municipality::new("Republic of Indexia".to_string()).get_or_create();
 
-        municipality.save();
-
-        let team = SportsTeam {
-            id: explicit.id,
-            name: "The Indexers".to_string(),
-            municipality: municipality.name,
-        };
-
-        team.save();
+        SportsTeam::new("The Indexers".to_string(), municipality.name).get_or_create();
     }
 
     fn fuel_indexer_test_panic(panic: Panic) {
@@ -567,13 +308,7 @@ mod fuel_indexer_test {
             reason,
         } = panic;
 
-        let panic = PanicEntity {
-            id: 123,
-            contract_id,
-            reason,
-        };
-
-        panic.save();
+        PanicEntity::new(contract_id, reason).get_or_create();
     }
 
     fn fuel_indexer_trigger_revert(revert: Revert) {
@@ -584,13 +319,7 @@ mod fuel_indexer_test {
             error_val,
         } = revert;
 
-        let entity = RevertEntity {
-            id: 123,
-            contract_id,
-            error_val,
-        };
-
-        entity.save();
+        RevertEntity::new(contract_id, error_val).get_or_create();
     }
 
     fn fuel_indexer_trigger_mint(mint: Mint) {
@@ -603,14 +332,7 @@ mod fuel_indexer_test {
             ..
         } = mint;
 
-        let entity = MintEntity {
-            id: 123,
-            sub_id,
-            contract_id,
-            val,
-        };
-
-        entity.save();
+        MintEntity::new(sub_id, contract_id, val).get_or_create();
     }
 
     fn fuel_indexer_trigger_burn(burn: Burn) {
@@ -623,14 +345,7 @@ mod fuel_indexer_test {
             ..
         } = burn;
 
-        let entity = BurnEntity {
-            id: 123,
-            sub_id,
-            contract_id,
-            val,
-        };
-
-        entity.save();
+        BurnEntity::new(sub_id, contract_id, val).get_or_create();
     }
 
     fn fuel_indexer_test_filterable_fields(_ping: Ping) {
@@ -696,13 +411,7 @@ mod fuel_indexer_test {
             error_val,
         } = enum_error;
 
-        let entity = EnumError {
-            id: 42,
-            contract_id,
-            error_val,
-        };
-
-        entity.save();
+        EnumError::new(contract_id, error_val).get_or_create();
     }
 
     fn fuel_indexer_block_explorer_types(_b: BlockData) {
@@ -720,11 +429,10 @@ mod fuel_indexer_test {
         e.save();
     }
 
-    #[allow(unused)]
     fn fuel_indexer_trigger_enum(
-        first: AnotherSimpleEnum,
-        second: NestedEnum,
-        third: AnotherSimpleEnum,
+        _first: AnotherSimpleEnum,
+        _second: NestedEnum,
+        _third: AnotherSimpleEnum,
     ) {
         info!("fuel_indexer_trigger_enum handling trigger_enum event..");
