+++ packages/fuel-indexer-tests/contracts/fuel-indexer-test/Forc.toml
@@ -5,4 +5,4 @@ license = "Apache-2.0"
 name = "fuel-indexer-test"
 
 [dependencies]
-std = { git = "https://github.com/FuelLabs/sway", branch = "upgrade/fuel-core-v0.20.1" }
+std = { git = "https://github.com/FuelLabs/sway", branch = "master" }
