+++ packages/fuel-indexer-tests/contracts/fuel-indexer-test/src/main.sw
@@ -182,7 +182,7 @@ impl FuelIndexer for Contract {
     #[payable]
     fn trigger_transferout() {
         const RECEIVER = Address::from(0x532ee5fb2cabec472409eb5f9b42b59644edb7bf9943eda9c2e3947305ed5e96);
-        transfer(Identity::Address(RECEIVER), BASE_ASSET_ID, 1);
+        transfer(msg_sender().unwrap(), BASE_ASSET_ID, 1);
     }
 
     #[payable]
