--- packages/fuel-indexer-tests/tests/e2e/mod.rs
@@ -1,5 +1,0 @@
-#[cfg(all(feature = "e2e", feature = "postgres"))]
-mod graphql_api_postgres;
-
-#[cfg(all(feature = "e2e", feature = "postgres"))]
-mod indexing_postgres;
