+++ packages/fuel-indexer-tests/tests/graphql/api.rs
@@ -1,19 +1,22 @@
 use fuel_indexer_tests::fixtures::{
-    http_client, mock_request, setup_web_test_components, WebTestComponents,
+    mock_request, setup_web_test_components, WebTestComponents,
 };
 use hyper::header::CONTENT_TYPE;
 use serde_json::{Number, Value};
 use std::collections::HashMap;
 
 #[actix_web::test]
-async fn test_can_return_query_response_with_all_fields_required_postgres() {
+async fn test_entity_with_required_and_optional_fields() {
     let WebTestComponents {
-        server, db: _db, ..
+        server,
+        db: _db,
+        client,
+        ..
     } = setup_web_test_components(None).await;
 
     mock_request("/block").await;
 
-    let client = http_client();
+    // All required
     let resp = client
         .post("http://127.0.0.1:29987/api/graph/fuel_indexer_test/index1")
         .header(CONTENT_TYPE, "application/graphql".to_owned())
@@ -32,18 +35,7 @@ async fn test_can_return_query_response_with_all_fields_required_postgres() {
     assert!(data[1]["height"].as_u64().unwrap() > 0);
     assert!(data[1]["timestamp"].as_u64().unwrap() > 0);
 
-    server.abort();
-}
-
-#[actix_web::test]
-async fn test_can_return_query_response_with_nullable_fields_postgres() {
-    let WebTestComponents {
-        server, db: _db, ..
-    } = setup_web_test_components(None).await;
-
-    mock_request("/optionals").await;
-
-    let client = http_client();
+    // Optionals
     let resp = client
         .post("http://127.0.0.1:29987/api/graph/fuel_indexer_test/index1")
         .header(CONTENT_TYPE, "application/graphql".to_owned())
@@ -64,14 +56,17 @@ async fn test_can_return_query_response_with_nullable_fields_postgres() {
 }
 
 #[actix_web::test]
-async fn test_can_return_nested_query_response_with_implicit_foreign_keys_postgres() {
+async fn test_entity_with_foreign_keys() {
     let WebTestComponents {
-        server, db: _db, ..
+        server,
+        db: _db,
+        client,
+        ..
     } = setup_web_test_components(None).await;
 
     mock_request("/block").await;
 
-    let client = http_client();
+    // Implicit foreign keys
     let resp = client
         .post("http://127.0.0.1:29987/api/graph/fuel_indexer_test/index1")
         .header(CONTENT_TYPE, "application/graphql".to_owned())
@@ -84,22 +79,44 @@ async fn test_can_return_nested_query_response_with_implicit_foreign_keys_postgr
     let v: Value = serde_json::from_str(&body).unwrap();
     let data = v["data"].as_array().expect("data is not an array");
 
-    assert!(data[0]["id"].as_i64().is_some());
-    assert!(data[0]["id"].as_i64().unwrap() > 0);
     assert!(data[0]["timestamp"].as_i64().is_some());
     assert!(data[0]["timestamp"].as_i64().unwrap() > 0);
-    assert!(data[0]["block"]["id"].as_i64().is_some());
-    assert!(data[0]["block"]["id"].as_i64().unwrap() > 0);
     assert!(data[0]["block"]["height"].as_i64().is_some());
     assert!(data[0]["block"]["height"].as_i64().unwrap() > 0);
+
+    // Explicit foreign keys
+    let resp = client
+        .post("http://127.0.0.1:29987/api/graph/fuel_indexer_test/index1")
+        .header(CONTENT_TYPE, "application/graphql".to_owned())
+        .body(
+            r#"{ "query": "query { sportsteam { id name municipality { id name } } }" }"#,
+        )
+        .send()
+        .await
+        .unwrap();
+
+    let body = resp.text().await.unwrap();
+    let v: Value = serde_json::from_str(&body).unwrap();
+    let data = v["data"].as_array().expect("data is not an array");
+
+    assert_eq!(data[0]["name"].as_str(), Some("The Indexers"));
+    assert!(data[0]["municipality"]["id"].as_i64().is_some());
+    assert!(data[0]["municipality"]["id"].as_i64().unwrap() > 0);
+    assert_eq!(
+        data[0]["municipality"]["name"].as_str(),
+        Some("Republic of Indexia")
+    );
 
     server.abort();
 }
 
 #[actix_web::test]
-async fn test_can_return_query_response_with_deeply_nested_query_postgres() {
+async fn test_deeply_nested_entity() {
     let WebTestComponents {
-        server, db: _db, ..
+        server,
+        db: _db,
+        client,
+        ..
     } = setup_web_test_components(None).await;
 
     mock_request("/deeply_nested").await;
@@ -165,7 +182,6 @@ async fn test_can_return_query_response_with_deeply_nested_query_postgres() {
             }",
     )]);
 
-    let client = http_client();
     let resp = client
         .post("http://127.0.0.1:29987/api/graph/fuel_indexer_test/index1")
         .header(CONTENT_TYPE, "application/graphql".to_owned())
@@ -230,48 +246,17 @@ async fn test_can_return_query_response_with_deeply_nested_query_postgres() {
 }
 
 #[actix_web::test]
-async fn test_can_return_nested_query_response_with_explicit_foreign_keys_postgres() {
+async fn test_filtering() {
     let WebTestComponents {
-        server, db: _db, ..
-    } = setup_web_test_components(None).await;
-
-    mock_request("/explicit").await;
-
-    let client = http_client();
-    let resp = client
-        .post("http://127.0.0.1:29987/api/graph/fuel_indexer_test/index1")
-        .header(CONTENT_TYPE, "application/graphql".to_owned())
-        .body(
-            r#"{ "query": "query { sportsteam { id name municipality { id name } } }" }"#,
-        )
-        .send()
-        .await
-        .unwrap();
-
-    let body = resp.text().await.unwrap();
-    let v: Value = serde_json::from_str(&body).unwrap();
-    let data = v["data"].as_array().expect("data is not an array");
-
-    assert_eq!(data[0]["name"].as_str(), Some("The Indexers"));
-    assert!(data[0]["municipality"]["id"].as_i64().is_some());
-    assert!(data[0]["municipality"]["id"].as_i64().unwrap() > 0);
-    assert_eq!(
-        data[0]["municipality"]["name"].as_str(),
-        Some("Republic of Indexia")
-    );
-
-    server.abort();
-}
-
-#[actix_web::test]
-async fn test_can_return_query_response_with_filter_id_selection_postgres() {
-    let WebTestComponents {
-        server, db: _db, ..
+        server,
+        db: _db,
+        client,
+        ..
     } = setup_web_test_components(None).await;
 
     mock_request("/ping").await;
 
-    let client = http_client();
+    // ID selection
     let resp = client
         .post("http://127.0.0.1:29987/api/graph/fuel_indexer_test/index1")
         .header(CONTENT_TYPE, "application/graphql".to_owned())
@@ -289,18 +274,7 @@ async fn test_can_return_query_response_with_filter_id_selection_postgres() {
     assert_eq!(data[0]["maybe_null_bar"].as_i64(), Some(123));
     assert_eq!(data[0]["bazoo"].as_i64(), Some(1));
 
-    server.abort();
-}
-
-#[actix_web::test]
-async fn test_can_return_query_response_with_filter_membership_postgres() {
-    let WebTestComponents {
-        server, db: _db, ..
-    } = setup_web_test_components(None).await;
-
-    mock_request("/ping").await;
-
-    let client = http_client();
+    // Set membership
     let resp = client
         .post("http://127.0.0.1:29987/api/graph/fuel_indexer_test/index1")
         .header(CONTENT_TYPE, "application/graphql".to_owned())
@@ -324,18 +298,7 @@ async fn test_can_return_query_response_with_filter_membership_postgres() {
     assert_eq!(data[1]["maybe_null_bar"].as_i64(), None);
     assert_eq!(data[1]["bazoo"].as_i64(), Some(5));
 
-    server.abort();
-}
-
-#[actix_web::test]
-async fn test_can_return_query_response_with_filter_non_null_postgres() {
-    let WebTestComponents {
-        server, db: _db, ..
-    } = setup_web_test_components(None).await;
-
-    mock_request("/ping").await;
-
-    let client = http_client();
+    // Non-null
     let resp = client
         .post("http://127.0.0.1:29987/api/graph/fuel_indexer_test/index1")
         .header(CONTENT_TYPE, "application/graphql".to_owned())
@@ -359,18 +322,7 @@ async fn test_can_return_query_response_with_filter_non_null_postgres() {
     assert_eq!(data[1]["maybe_null_bar"].as_i64(), Some(456));
     assert_eq!(data[1]["bazoo"].as_i64(), Some(1000));
 
-    server.abort();
-}
-
-#[actix_web::test]
-async fn test_can_return_query_response_with_filter_complex_comparison_postgres() {
-    let WebTestComponents {
-        server, db: _db, ..
-    } = setup_web_test_components(None).await;
-
-    mock_request("/ping").await;
-
-    let client = http_client();
+    // Complex comparison
     let resp = client
         .post("http://127.0.0.1:29987/api/graph/fuel_indexer_test/index1")
         .header(CONTENT_TYPE, "application/graphql".to_owned())
@@ -394,17 +346,7 @@ async fn test_can_return_query_response_with_filter_complex_comparison_postgres(
     assert_eq!(data[1]["maybe_null_bar"].as_i64(), None);
     assert_eq!(data[1]["bazoo"].as_i64(), Some(5));
 
-    server.abort();
-}
-
-#[actix_web::test]
-async fn test_can_return_query_response_with_filter_simple_comparison_postgres() {
-    let WebTestComponents {
-        server, db: _db, ..
-    } = setup_web_test_components(None).await;
-
-    mock_request("/ping").await;
-    let client = http_client();
+    // Simple comparison
     let resp = client
         .post("http://127.0.0.1:29987/api/graph/fuel_indexer_test/index1")
         .header(CONTENT_TYPE, "application/graphql".to_owned())
@@ -426,18 +368,7 @@ async fn test_can_return_query_response_with_filter_simple_comparison_postgres()
     assert_eq!(data[1]["maybe_null_bar"].as_i64(), None);
     assert_eq!(data[1]["bazoo"].as_i64(), Some(5));
 
-    server.abort();
-}
-
-#[actix_web::test]
-async fn test_can_return_query_response_with_filter_nested_postgres() {
-    let WebTestComponents {
-        server, db: _db, ..
-    } = setup_web_test_components(None).await;
-
-    mock_request("/ping").await;
-
-    let client = http_client();
+    // Nested filters
     let resp = client
         .post("http://127.0.0.1:29987/api/graph/fuel_indexer_test/index1")
         .header(CONTENT_TYPE, "application/graphql".to_owned())
@@ -461,18 +392,7 @@ async fn test_can_return_query_response_with_filter_nested_postgres() {
     assert_eq!(data[0]["inner_entity"]["inner_bar"].as_u64(), Some(500));
     assert_eq!(data[0]["inner_entity"]["inner_baz"].as_u64(), Some(600));
 
-    server.abort();
-}
-
-#[actix_web::test]
-async fn test_can_return_query_response_with_filter_multiple_on_single_entity_postgres() {
-    let WebTestComponents {
-        server, db: _db, ..
-    } = setup_web_test_components(None).await;
-
-    mock_request("/ping").await;
-
-    let client = http_client();
+    // Multiple filters on single entity
     let resp = client
         .post("http://127.0.0.1:29987/api/graph/fuel_indexer_test/index1")
         .header(CONTENT_TYPE, "application/graphql".to_owned())
@@ -492,18 +412,7 @@ async fn test_can_return_query_response_with_filter_multiple_on_single_entity_po
     assert_eq!(data[0]["maybe_null_bar"].as_i64(), Some(123));
     assert_eq!(data[0]["bazoo"].as_i64(), Some(1));
 
-    server.abort();
-}
-
-#[actix_web::test]
-async fn test_can_return_query_response_with_filter_negation_postgres() {
-    let WebTestComponents {
-        server, db: _db, ..
-    } = setup_web_test_components(None).await;
-
-    mock_request("/ping").await;
-
-    let client = http_client();
+    // Negation
     let resp = client
         .post("http://127.0.0.1:29987/api/graph/fuel_indexer_test/index1")
         .header(CONTENT_TYPE, "application/graphql".to_owned())
@@ -527,13 +436,16 @@ async fn test_can_return_query_response_with_filter_negation_postgres() {
 }
 
 #[actix_web::test]
-async fn test_can_return_query_response_with_sorted_results_postgres() {
+async fn test_sorting() {
     let WebTestComponents {
-        server, db: _db, ..
+        server,
+        db: _db,
+        client,
+        ..
     } = setup_web_test_components(None).await;
 
     mock_request("/ping").await;
-    let client = http_client();
+
     let resp = client
         .post("http://127.0.0.1:29987/api/graph/fuel_indexer_test/index1")
         .header(CONTENT_TYPE, "application/graphql".to_owned())
@@ -559,15 +471,16 @@ async fn test_can_return_query_response_with_sorted_results_postgres() {
 }
 
 #[actix_web::test]
-async fn test_can_return_query_response_with_alias_and_ascending_offset_and_limited_results_postgres(
-) {
+async fn test_aliasing_and_pagination() {
     let WebTestComponents {
-        server, db: _db, ..
+        server,
+        db: _db,
+        client,
+        ..
     } = setup_web_test_components(None).await;
 
     mock_request("/ping").await;
 
-    let client = http_client();
     let resp = client
         .post("http://127.0.0.1:29987/api/graph/fuel_indexer_test/index1")
         .header(CONTENT_TYPE, "application/graphql".to_owned())
@@ -582,7 +495,6 @@ async fn test_can_return_query_response_with_alias_and_ascending_offset_and_limi
     let v: Value = serde_json::from_str(&body).unwrap();
     let data = v["data"].as_array().expect("data is not an array");
 
-    assert_eq!(data[0]["aliased_entities"][0]["id"].as_i64(), Some(3));
     assert_eq!(
         data[0]["aliased_entities"][0]["foola"].as_str(),
         Some("blorp")
