+++ packages/fuel-indexer-tests/tests/graphql/mod.rs
@@ -1,0 +1,2 @@
+mod api;
+mod schema;
