+++ packages/fuel-indexer-tests/tests/graphql/schema.rs
