--- packages/fuel-indexer-tests/tests/integration/mod.rs
@@ -1,4 +1,0 @@
-mod database;
-mod graphql_schema;
-mod service;
-mod web_api_postgres;
