+++ packages/fuel-indexer-tests/tests/wasm/mod.rs
@@ -1,0 +1,2 @@
+mod database;
+mod service;
